fix routing landing page
